consolidate files
