+++ cmd/workflow/deploy/registry_adapter.go
@@ -2,8 +2,9 @@ package deploy
 
 import (
 	"errors"
+	"fmt"
+	"strings"
 
-	"github.com/smartcontractkit/cre-cli/cmd/client"
 	"github.com/smartcontractkit/cre-cli/internal/environments"
 )
 
@@ -26,10 +27,47 @@ type registryAdapter interface {
 	Upsert() error
 }
 
+type registryTargetType string
+
+const (
+	registryTargetOnchain registryTargetType = "onchain"
+	registryTargetPrivate registryTargetType = "private"
+)
+
+type registryTarget struct {
+	targetType registryTargetType
+}
+
+func (t registryTarget) isPrivate() bool {
+	return t.targetType == registryTargetPrivate
+}
+
+// resolveRegistryTarget determines the target workflow registry from inputs and
+// environment. Preview-specific naming (--preview-private-registry, STAGING
+// gate) is confined to this function; the returned target is preview-agnostic.
+func resolveRegistryTarget(previewPrivateRegistry bool, envSet *environments.EnvironmentSet) (registryTarget, error) {
+	if previewPrivateRegistry {
+		if err := validatePrivateRegistryAllowed(envSet); err != nil {
+			return registryTarget{}, err
+		}
+		return registryTarget{targetType: registryTargetPrivate}, nil
+	}
+
+	return registryTarget{targetType: registryTargetOnchain}, nil
+}
+
+// validatePrivateRegistryAllowed enforces the STAGING-only preview gate.
+func validatePrivateRegistryAllowed(envSet *environments.EnvironmentSet) error {
+	if !strings.EqualFold(envSet.EnvName, "STAGING") {
+		return fmt.Errorf("--preview-private-registry is only available in the STAGING environment (current: %s)", envSet.EnvName)
+	}
+	return nil
+}
+
 // resolveTargetRegistry resolves the deploy target details and returns the
 // adapter that handles target-specific registry operations.
-func resolveTargetRegistry(previewPrivateRegistry bool, envSet *environments.EnvironmentSet, factory client.Factory, h *handler) (registryTarget, registryAdapter, error) {
-	target, err := resolveRegistryTarget(previewPrivateRegistry, envSet, factory)
+func resolveTargetRegistry(previewPrivateRegistry bool, envSet *environments.EnvironmentSet, h *handler) (registryTarget, registryAdapter, error) {
+	target, err := resolveRegistryTarget(previewPrivateRegistry, envSet)
 	if err != nil {
 		return registryTarget{}, nil, err
 	}
