+++ cmd/workflow/deploy/registry_target_test.go
@@ -6,25 +6,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/smartcontractkit/cre-cli/cmd/client"
 	"github.com/smartcontractkit/cre-cli/internal/environments"
 )
-
-type fakeFactory struct {
-	txType client.TxType
-}
-
-func (f fakeFactory) NewWorkflowRegistryV2Client() (*client.WorkflowRegistryV2Client, error) {
-	return nil, nil
-}
-
-func (f fakeFactory) GetTxType() client.TxType {
-	return f.txType
-}
-
-func (f fakeFactory) GetSkipConfirmation() bool {
-	return false
-}
 
 func TestResolveRegistryTarget(t *testing.T) {
 	t.Parallel()
@@ -36,13 +19,10 @@ func TestResolveRegistryTarget(t *testing.T) {
 			WorkflowRegistryChainName: "ethereum-testnet-sepolia",
 			WorkflowRegistryAddress:   "0x1234567890123456789012345678901234567890",
 		}
-		factory := fakeFactory{txType: client.Regular}
 
-		target, err := resolveRegistryTarget(false, envSet, factory)
+		target, err := resolveRegistryTarget(false, envSet)
 		require.NoError(t, err)
 		assert.Equal(t, registryTargetOnchain, target.targetType)
-		assert.NotZero(t, target.onchainChainSelector)
-		assert.Equal(t, "0x1234567890123456789012345678901234567890", target.onchainAddress)
 		assert.False(t, target.isPrivate())
 	})
 
@@ -53,9 +33,8 @@ func TestResolveRegistryTarget(t *testing.T) {
 			WorkflowRegistryChainName: "ethereum-testnet-sepolia",
 			WorkflowRegistryAddress:   "0x1234567890123456789012345678901234567890",
 		}
-		factory := fakeFactory{txType: client.Regular}
 
-		target, err := resolveRegistryTarget(true, envSet, factory)
+		target, err := resolveRegistryTarget(true, envSet)
 		require.NoError(t, err)
 		assert.Equal(t, registryTargetPrivate, target.targetType)
 		assert.True(t, target.isPrivate())
@@ -68,9 +47,8 @@ func TestResolveRegistryTarget(t *testing.T) {
 			WorkflowRegistryChainName: "ethereum-testnet-sepolia",
 			WorkflowRegistryAddress:   "0x1234567890123456789012345678901234567890",
 		}
-		factory := fakeFactory{txType: client.Regular}
 
-		_, err := resolveRegistryTarget(true, envSet, factory)
+		_, err := resolveRegistryTarget(true, envSet)
 		require.Error(t, err)
 		assert.Contains(t, err.Error(), "--preview-private-registry is only available in the STAGING environment")
 	})
@@ -82,9 +60,8 @@ func TestResolveRegistryTarget(t *testing.T) {
 			WorkflowRegistryChainName: "ethereum-testnet-sepolia",
 			WorkflowRegistryAddress:   "0x1234567890123456789012345678901234567890",
 		}
-		factory := fakeFactory{txType: client.Regular}
 
-		_, err := resolveRegistryTarget(true, envSet, factory)
+		_, err := resolveRegistryTarget(true, envSet)
 		require.Error(t, err)
 		assert.Contains(t, err.Error(), "--preview-private-registry is only available in the STAGING environment")
 	})
@@ -96,9 +73,8 @@ func TestResolveRegistryTarget(t *testing.T) {
 			WorkflowRegistryChainName: "ethereum-testnet-sepolia",
 			WorkflowRegistryAddress:   "0x1234567890123456789012345678901234567890",
 		}
-		factory := fakeFactory{txType: client.Regular}
 
-		target, err := resolveRegistryTarget(true, envSet, factory)
+		target, err := resolveRegistryTarget(true, envSet)
 		require.NoError(t, err)
 		assert.True(t, target.isPrivate())
 	})
